Test-Driven Development and GUI Programming
In this first part of the assignment you will complete the
implementation of some Java packages that allow you to maintain a
ship’s cargo manifest, i.e., a record of the cargo containers currently
on board. (In Part 2 of the assignment you will provide a Graphical
User Interface to allow the ship’s captain to manipulate the manifest.)

The scenario on which this assignment is based was described in some
detail in the specification for Part 1. Here we shall be very focused
on the development of a simple GUI to support interaction with the API.
The GUI which we required of you for the completion of this assignment
need not be terribly sophisticated, but it must be functional and it
must not be excessively cluttered. As noted earlier, you may use a GUI
builder as long as you declare this in your statement of completeness
(See forthcoming submission instructions).